# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Logenhelfer API
  version: 1.0.0
  description: Backend-agnostic contract. Rails serves it today, Express later.
paths:
...
  /api/v1/app_config:
    get:
      operationId: getAppConfig
      responses:
        '200':
          description: Every known AppConfig setting and its current value
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppConfigValues' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      operationId: updateAppConfig
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppConfigValues' }
      responses:
        '200':
          description: The full, updated set of AppConfig settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppConfigValues' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
...
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
...
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            required: [error, detail]
            properties:
              error: { type: string }
              detail: { type: string }
  schemas:
...
    AppConfigValues:
      type: object
      additionalProperties: false
      properties:
        public_wp_available_to_anon_users: { type: boolean, nullable: true }
        working_plan_as_start_page: { type: boolean, nullable: true }
        archive: { type: boolean, nullable: true }
        show_admins: { type: boolean, nullable: true }
        users_can_view_statistics: { type: boolean, nullable: true }
        show_seeker_names_to_brothers: { type: boolean, nullable: true }
        notify_technical_contact_on_unknown_password_reset: { type: boolean, nullable: true }
        domain: { type: string, nullable: true }
        organisation: { type: string, nullable: true }
        lodge: { type: string, nullable: true }
        lodge_short: { type: string, nullable: true }
        language: { type: string, nullable: true }
        default_workingplan_timespan: { type: integer, nullable: true }
        default_event_location: { type: string, nullable: true }
        default_event_duration_minutes: { type: integer, nullable: true }
        user_change_notification_email: { type: string, nullable: true }
        default_from_email: { type: string, nullable: true }
        technical_contact_email: { type: string, nullable: true }
        mvst_email: { type: string, nullable: true }
        zip: { type: string, nullable: true }
        location: { type: string, nullable: true }
        max_db_mem_size: { type: string, nullable: true }
        max_upload_file_size: { type: string, nullable: true }
        workingplan_footer: { type: string, nullable: true }
        impressum: { type: string, nullable: true }
        help: { type: string, nullable: true }
        robots_txt: { type: string, nullable: true }
        public_workingplan_html_timespan: { type: integer, nullable: true }
        public_workingplan_ics_timespan: { type: integer, nullable: true }
        mfa_mode: { type: string, nullable: true }
        mfa_enforce_for_officers: { type: boolean, nullable: true }
        mfa_grace_period_days: { type: integer, nullable: true }
        mfa_trusted_device_days: { type: integer, nullable: true }
security:
  - bearerAuth: []